Check hotkey permutation consistency within the loop
This feels cleaner to me.  And as for potentially gathering all the
inconsistent noaborts and duplicate permutations, I'll do it if it's
needed.

It might be good to indicate all the sets of duplicate permutations.
However, I'm not so sure how good error messages could be made for the
inconsistent noabort case.