hkxargs: Run --exec/-e commands in the background
Some commands blocked, meaning that for my usecase, macros, sometimes
macros stopped in the middle of execution.

I'm not exactly sure what sxhkd means for a command to be started
synchronously, but I assume it means that the command must be completed
before any others are started.

Also need a delay to allow for Xorg etc. updates to be done on time.