lib: Store Command.permutations as List[str] again
It makes most sense for the `permutations` attribute of each type to be
the one that makes most sense: chord chains for hotkeys and strings for
commands.

If one wants the raw `List[SpanPermutation]`, it can be easily obtained
using the public `span_tree` attribute and then calling
`generate_permutations()` on it.