hkwatch: Remove useless assertion
It also wasn't correct, since the order of internal nodes is
significant--so the set comparisons throw away information.  It was
mostly there to guard against passing trees that break the assumptions
of the function, but it's not much of a problem because the internal
nodes come from one source anyway (the `INTERNAL_NODES` constant).