lib: Store noabort information outside of chords
As part of this, I also made some auxiliary changes I've been wanting to
do--namely to refine the data model so that invalid states are fewer and
further between.

Having noabort information stored outside of chords has some pros and
cons, but the main reason for doing this is because it deduplicates the
noabort information--now just on the `HotkeyPermutation` and the
`HotkeyTreeChordData` objects.

As a consequence, it's very easy to change the noabort index for a
permutation and there is no danger of desynchronisation between the
permutation and what the sequence of chords actually has in terms of
noaborts.

However, it's a bit more fiddly to do a tree search since there has to
be extra logic added to take the separate noabort information into
account.  The benefits outlined above seem worth it though.