cli: Avoid using `argparse.ArgumentDefaultsHelpFormatter`
Adding the defaults to the help string of each option that needs it is
better than having useless and awkward-looking defaults for those that
don't need it.

The section/metadata config options other than the `--*-type` options
don't need defaults in the help string.  Their values are indeterminate
as far as the user is concerned.

Resolves #2